pyproject.toml
[project]
name = "kurtis"
version = "0.1.2"
description = "Kurtis SLM Toolkit"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Massimo R. Scamarcia", email = "massimo.scamarcia@gmail.com"},
]
dependencies = [
    "chardet>=5.2.0",
    "click>=8.1.7",
    "einops>=0.8.0",
    "nltk>=3.9.1",
    "prompt-toolkit>=3.0.48",
    "protobuf>=5.28.3",
    "pyfiglet>=1.0.2",
    "rich>=13.9.4",
    "rouge-score>=0.1.2",
    "safetensors>=0.4.5",
    "scikit-learn>=1.5.2",
    "sentencepiece>=0.2.0",
    "urwid>=2.6.16",
    # torch dependencies managed via uv.sources and optional-dependencies
    "datasets>=3.1.0",
    "numpy>=2.1.3",
    "tqdm>=4.67.0",
    "setuptools>=75.3.0",
    "openai>=1.76.0",
    "python-slugify>=8.0.4",
    "huggingface-hub[hf-xet]>=0.30.2",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
gpu = []

[project.scripts]
kurtis = "kurtis.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kurtis"]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '^/data/'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
skip_glob = ["data/*"]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["data"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "T20", # flake8-print
]
ignore = [
    "E501", # Ignore line length errors (let black handle it)
    "B008", # Allow function calls in defaults (FastAPI Depends)
    "E402", # Allow module level imports not at top (dynamic imports)
    "T201", # Allow print statements
]

[tool.ruff.lint.isort]
known-first-party = ["kurtis"]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=7.0.0",
]
rocm = [
    "torch",
    "torchaudio",
    "torchvision",
    "pytorch-triton-rocm",
    "peft",
    "transformers",
    "trl",
    "evaluate",
    "accelerate",
    "bitsandbytes>=0.46.1",
]

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/nightly/rocm7.1"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-rocm" }
pytorch-triton-rocm = { index = "pytorch-rocm" }
torchaudio = { index = "pytorch-rocm" }
torchvision = { index = "pytorch-rocm" }


[tool.uv]
prerelease = "allow"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestRemovedIn9Warning",
]
